created new SamRequestContextDirectives and unfortunately copied the private trace directives from TraceDirectives. The other option would be to change the accessibility (which is modeled in a prior commit) but may run counter to SecurityManager.
