+++ src/main/scala/org/broadinstitute/dsde/workbench/sam/api/UserRoutes.scala
@@ -15,7 +15,7 @@ import scala.concurrent.ExecutionContext
 /**
   * Created by mbemis on 5/22/17.
   */
-trait UserRoutes extends UserInfoDirectives with SamModelDirectives {
+trait UserRoutes extends UserInfoDirectives with SamRequestContextDirectives {
   implicit val executionContext: ExecutionContext
   val userService: UserService
